Add interface fields to getNodes
Previously we were not traversing interface fields in the generated
getNodes implementations, while we should.

It wasn't quite as simple as adjusting the condition though, because
while all interfaces are implemented by concrete types extending from
AbstractResponse, the java type system cannot represent the concept of
an interface extending a class. Add a self() method to AbstractResponse
and put it in each generated interface to solve.

I suppose an alternative might be to make AbstractResponse an interface
itself and then have a ConcreteAbstractResponse implements
AbstractResponse or something like that.